1.0.5

+++ plugin/src/main/java/com/dailyyoga/plugin/net/InjectJar.java
@@ -82,6 +82,9 @@ public abstract class InjectJar {
         try {
             if (mOriginFile == null || mTempFile == null || mDestFile == null) return;
             processJar();
+            if (mTempFile.getCanonicalPath().equals(mDestFile.getCanonicalPath())) {
+                return;
+            }
             FileUtils.copyFile(mTempFile, mDestFile);
             mTempFile.delete();
         } catch (IOException e) {
